build: clean the build dir on each run
This should allow to switch easier between targets [now that it's
supported].
We should be pretty confident [by now] that buildroot, linux, hdl and
u-boot can do similar switching between builds (and target artifacts).

Also, to do this properly we need to enforce the build to not run in
parallel if the -jN option is provided.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>