scripts/legal_info_html: Fix HTTP 301 Moved Permanently issue with libxml2
The script checks for thw presence of the download sources.
When checking libxml2 ,HTTP 301 is returned with a relative location/path
on the same host. In fact it just complains about a missing '/' in the
URL. Assuming the location doesn't contain a protocol stop and validate
the original URL.


$ curl -IsS https://download.gnome.org/sources/libxml2/2.10
HTTP/2 301
server: nginx/1.24.0
content-type: text/html
location: /sources/libxml2/2.10/


$ curl -IsS https://download.gnome.org/sources/libxml2/2.10/
HTTP/2 200

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>